%% File hydroGOF-internal.Rd
%% Part of the hydroGOF R package, https://github.com/hzambran/hydroGOF ; 
%%                                 https://cran.r-project.org/package=hydroGOF
%%                                 http://www.rforge.net/hydroGOF/
%% Copyright 2008-2024 Mauricio Zambrano-Bigiarini
%% Distributed under GPL 2 or later

\name{hydroGOF-internal}
\title{Internal hydroGOF objects}
\alias{.intersect}
\alias{preproc}
\description{Internal hydroGOF objects.}
\details{These are not to be called by the user.}
\keyword{internal}
