%% File EgaEnEstellaQts.Rd
%% Part of the hydroGOF R package, http://www.rforge.net/hydroGOF/ ; 
%%                                 http://cran.r-project.org/web/packages/hydroGOF/
%% Copyright 2011-2017 Mauricio Zambrano-Bigiarini
%% Distributed under GPL 2 or lat

\name{EgaEnEstellaQts}
\docType{data}
\alias{EgaEnEstellaQts}
\title{
Ega in "Estella" (Q071), ts with daily streamflows.
}
\description{
Time series with daily streamflows of the Ega River (subcatchment of the Ebro River basin, Spain) measured at the gauging station "Estella" (Q071), for the period 01/Jan/1961 to 31/Dec/1970
}
\usage{
data(EgaEnEstellaQts)
}

%%\details{
%%
%%}
\format{
zoo object.
}
\source{
Downloaded from: \url{http://www.chebro.es}. Last accessed [March 2010]. \cr
These data are intended to be used for research purposes only, being distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY.
}

%%\references{
%% ~put references to the literature/web site here ~
%%}
\keyword{datasets}
